incomplete update from Omer and Isaac
If the contract does not have a parameter address, the SENDER address should be included as the @Originator address in storage, when the map is empty. Later, when a parameter is included, a check is made to see if the SENDER is the same as the @Originator, and if the tz@parameter is not already in the map. if not in the map it is included